WIP Send messages from TView ui
